Mip heuristic (#100)
* Fix some bugs on when stopping the branching.

* Add a mip heuristic to solve the problem with current columns. Use the BCP feature granulaity for the optamility gap. Continue to investigate the bug about the linear relaxation objective.

* Transform all parameters to upper cases when parsing params.

* Fix potential bad access errors in BcpHeuristic (some deleted objects could be reused.). Print some information to track the relaxation bug.

* cpplint code. fix some details in python script and add some retries in case an action fails. display infinity instead fo LARGE_SCORE for summary lines.

* Revert change on LARGE_SCORE value. Otherwise, it is not big enough.

* Change management of PThreadsPool to be deleted at the right time.

* Add a retries option to docker-entrypoint.sh.

* Set retries to 5 for docker tests.

* Add flags to enable heuristics.

+++ src/tools/InputPaths.h
@@ -31,7 +31,7 @@ class InputPaths {
              const std::string &solutionPath = "",
              const std::string &logPath = "",
              const std::string &paramFile = "",
-             double timeOut = -1,
+             int timeOut = -1,
              int verbose = -1,
              int randSeed = 0,
              const std::string &SPType = "",
@@ -46,7 +46,7 @@ class InputPaths {
              const std::string &solutionPath = "",
              const std::string &logPath = "",
              const std::string &paramFile = "",
-             double timeOut = -1,
+             int timeOut = -1,
              int verbose = -1,
              int randSeed = 0,
              const std::string &SPType = "",
@@ -69,7 +69,7 @@ class InputPaths {
   std::string paramFile_ = "";
   int verbose_ = -1;
   int randSeed_ = 0;
-  double timeOut_ = 3600;
+  int timeOut_ = 3600;
   std::string SPType_ = "LONG";
   int SPStrategy_ = 0;
   std::string RCSPPType_ = "BOOST";
@@ -104,8 +104,8 @@ class InputPaths {
 
   int randSeed() const { return randSeed_; }
   void randSeed(int seed) { randSeed_ = seed; }
-  double timeOut() const { return timeOut_; }
-  void timeOut(double t) { timeOut_ = t; }
+  int timeOut() const { return timeOut_; }
+  void timeOut(int t) { timeOut_ = t; }
 
   int verbose() const { return verbose_; }
   void verbose(int verbose) { verbose_ = verbose; }
